refactor(cli): split argument helpers
Move output format, render option, input source, and error helpers out of the CLI coordinator so argument handling remains easier to review without changing the command surface.
