+++ src/cli/errors.ts
@@ -1,0 +1,5 @@
+export function exitWithError(message: string): never {
+  console.error(`Error: ${message}`);
+  console.error('Run "pdfvision --help" for usage.');
+  process.exit(1);
+}
